signed/unsigned

+++ init/languages.cpp
@@ -146,7 +146,7 @@ void edit_languages() {
         }
 
         string prompt = StringPrintf("Insert before which (1-%d) : ", languages.size() + 1);
-        int i = dialog_input_number(out->window(), prompt, 1, languages.size() + 1);
+        size_t i = dialog_input_number(out->window(), prompt, 1, languages.size() + 1);
         // N.B. i is one based, result.selected is 0 based.
         if (i <= 0 || i > languages.size() + 1) {
           break;
